pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-server-odoo"
version = "2.0.0"
description = "Enhanced MCP server for Odoo with improved authentication and LLM optimization"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0"}
authors = [
    {name = "Enhanced by Replit Agent"}
]
keywords = ["mcp", "odoo", "erp", "ai", "llm", "n8n", "claude", "chatgpt"]

dependencies = [
    "mcp>=0.9.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "redis>=5.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
mcp-server-odoo = "mcp_server_odoo.__main__:main"

[tool.setuptools.packages.find]
include = ["mcp_server_odoo*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
